switched register and login page.
