+++ src/app/hooks/useRequireAuth.tsx
@@ -4,7 +4,7 @@ import { useEffect } from 'react';
 import { useRouter } from 'next/navigation';
 import { useAuth } from '@/contexts/auth-context-provider';
 
-const useRequireAuth = (redirectTo: string = '/login') => {
+const useRequireAuth = (redirectTo: string = '/') => {
   const { isLoggedIn } = useAuth();
   const router = useRouter();
